add hash, and Verifier, cleanup DNSKEY usage
