+++ CHANGELOG.md
@@ -7,6 +7,14 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Fixed format of ED25519 keys (@briansmith) #129
 
+### Changed
+- Revamped signer and keypair to better deal with public key
+- Upgraded *ring* to 0.9.x series, requires pkcs8 for key storage
+
+### Added
+- PublicKey and Verifier for verifying with zero copy from KEY and DNSKEY
+- Pkcs8 as a supported KeyFormat for storage
+
 ## 0.10.1
 ### Added
 - Added `From<IpAddr>` for Name (reverse DNS) #105
