+++ client/src/rr/dnssec/mod.rs
@@ -18,14 +18,16 @@
 
 mod algorithm;
 mod digest_type;
+pub mod hash;
 #[cfg(any(feature = "openssl", feature = "ring"))]
 mod key_format;
 mod keypair;
 mod nsec3;
-mod public_key;
+pub mod public_key;
 mod signer;
 mod supported_algorithm;
 mod trust_anchor;
+mod verifier;
 
 pub use self::algorithm::Algorithm;
 pub use self::digest_type::DigestType;
@@ -38,6 +40,7 @@ pub use self::public_key::PublicKeyEnum;
 pub use self::signer::Signer;
 pub use self::supported_algorithm::SupportedAlgorithms;
 pub use self::trust_anchor::TrustAnchor;
+pub use self::verifier::Verifier;
 
 pub use error::DnsSecError;
 pub use error::DnsSecErrorKind;
